requirements.txt
agate==1.6.3
apache-beam==2.37.0
appnope==0.1.3
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
crcmod==1.7
cryptography==42.0.5
cycler==0.11.0
db-dtypes==1.2.0
dbt-bigquery==1.0.0
dbt-core==1.0.3
dbt-extractor==0.4.0
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.4
docopt==0.6.2
entrypoints==0.4
fastavro==1.4.10
fasteners==0.17.3
filelock==3.6.0
fonttools==4.38.0
future==0.18.2
fuzzywuzzy==0.18.0
google-api-core==2.17.1
google-apitools==0.5.31
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-bigquery==3.19.0
google-cloud-bigquery-storage==2.24.0
google-cloud-bigtable==1.7.1
google-cloud-core==2.3.0
google-cloud-datastore==1.15.4
google-cloud-dlp==3.6.2
google-cloud-language==1.3.1
google-cloud-pubsub==2.12.0
google-cloud-pubsublite==1.4.1
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.2
google-cloud-storage==2.3.0
google-cloud-videointelligence==1.16.2
google-cloud-vision==1.0.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-gcp==0.2.2
grpcio-status==1.44.0
hdfs==2.7.0
hologram==0.0.14
httplib2==0.19.1
idna==3.3
importlib-metadata==4.11.2
ipykernel==6.13.0
ipython==7.32.0
isodate==0.6.1
jedi==0.18.1
Jinja2==2.11.3
joblib==1.3.2
jsonschema==3.1.1
jupyter-client==7.2.2
jupyter-core==4.10.0
kiwisolver==1.4.5
leather==0.3.4
Levenshtein==0.23.0
Logbook==1.5.3
lxml==5.1.0
MarkupSafe==2.0.1
mashumaro==2.9
matplotlib==3.5.3
matplotlib-inline==0.1.3
minimal-snowplow-tracker==0.0.2
msgpack==1.0.3
nest-asyncio==1.5.5
networkx==2.6.3
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.6.8
overrides==6.1.0
packaging==21.3
pandas==1.3.5
pandas-gbq==0.21.0
parsedatetime==2.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2022.4.8
platformdirs==2.5.1
prompt-toolkit==3.0.29
proto-plus==1.23.0
protobuf==4.24.4
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybaseball @ git+https://github.com/johntylernyc/pybaseball.git@1c051622d83360958e4cf368069f734f103f90ea
pycparser==2.21
pydata-google-auth==1.8.2
pydot==1.4.2
PyGithub==2.2.0
Pygments==2.11.2
PyJWT==2.8.0
pymongo==3.12.3
PyNaCl==1.5.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-Levenshtein==0.23.0
python-slugify==6.1.1
pytimeparse==1.1.8
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
rapidfuzz==3.4.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.2
text-unidecode==1.3
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.66.2
traitlets==5.1.1
typing-utils==0.1.0
typing_extensions==4.7.1
urllib3==1.26.8
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.16.0
zipp==3.7.0
